// src/common/exceptions/error-code.enum.ts
export enum ErrorCode {
  VALIDATION_MISSING_FIELD = 'VALIDATION_MISSING_FIELD',
  VALIDATION_CUSTOMER_TAXID_EXISTS = 'VALIDATION_CUSTOMER_TAXID_EXISTS',
  VALIDATION_CUSTOMER_PASSPORT_EXISTS = 'VALIDATION_CUSTOMER_PASSPORT_EXISTS',
  VALIDATION_CUSTOMER_HAS_ORDERS = 'VALIDATION_CUSTOMER_HAS_ORDERS',
  VALIDATION_INVALID_CUSTOMER_ID = 'VALIDATION_INVALID_CUSTOMER_ID',
  VALIDATION_ORDER_NOT_EDITABLE = 'VALIDATION_ORDER_NOT_EDITABLE',
  VALIDATION_ORDER_EMPTY = 'VALIDATION_ORDER_EMPTY',
  VALIDATION_ORDER_INVALID_STATUS_TRANSITION = 'VALIDATION_ORDER_INVALID_STATUS_TRANSITION',
  VALIDATION_INVALID_FILTER_VALUE = 'VALIDATION_INVALID_FILTER_VALUE',
  VALIDATION_INVALID_SORT_FIELD = 'VALIDATION_INVALID_SORT_FIELD',
  VALIDATION_CONSTRAINT_VIOLATION = 'VALIDATION_CONSTRAINT_VIOLATION',

  RESOURCE_NOT_FOUND_CUSTOMER = 'RESOURCE_NOT_FOUND_CUSTOMER',
  RESOURCE_NOT_FOUND_ORDER = 'RESOURCE_NOT_FOUND_ORDER',
  RESOURCE_NOT_FOUND_ITEM = 'RESOURCE_NOT_FOUND_ITEM',

  CONFLICT_CONCURRENT_MODIFICATION = 'CONFLICT_CONCURRENT_MODIFICATION',
  CONFLICT_DATA_INTEGRITY_VIOLATION = 'CONFLICT_DATA_INTEGRITY_VIOLATION',

  AUTHORIZATION_ACCESS_DENIED = 'AUTHORIZATION_ACCESS_DENIED',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
}